+++ src/main/java/apimining/upminer/UPMiner.java
@@ -1,8 +1,10 @@
 package apimining.upminer;
 
+import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.io.InputStreamReader;
 import java.io.PrintWriter;
 import java.util.Collection;
 import java.util.Map.Entry;
@@ -19,12 +21,12 @@ public class UPMiner {
 
 	public static void main(final String[] args) throws Exception {
 
-		final String project = "andengine";
+		final String project = "cloud9";
 		final String arffFile = "/afs/inf.ed.ac.uk/user/j/jfowkes/Code/Sequences/Datasets/API/examples/calls/" + project
 				+ ".arff";
 		final String outFolder = "/afs/inf.ed.ac.uk/user/j/jfowkes/Code/Sequences/Datasets/API/examples/" + project
 				+ "/upminer/";
-		mineAPICallSequences(arffFile, outFolder, 1, 1, 0.3);
+		mineAPICallSequences(arffFile, outFolder, 1, 1);
 
 	}
 
@@ -36,52 +38,67 @@ public class UPMiner {
 	 *            as space separated string of API calls.
 	 */
 	public static void mineAPICallSequences(final String arffFile, final String outFolder, final int noClusters1,
-			final int noClusters2, final double minSupp) throws Exception {
+			final int noClusters2) throws Exception {
 
 		new File(outFolder).mkdirs();
-		final File arffFileFreq = File.createTempFile("FreqCalls", ".arff");
-		writeArffHeader(arffFileFreq);
 
 		System.out.print("===== Clustering call sequences #1... ");
 		final Multimap<Integer, String> clusteredCallSeqs1 = APICallClustererSequence.clusterAPICallSeqs(arffFile,
 				noClusters1);
 		System.out.println("done. Number of clusters: " + clusteredCallSeqs1.keySet());
 
-		int count = 0;
-		for (final Collection<String> callSeqs : clusteredCallSeqs1.asMap().values()) {
+		double minSupp = 0.3;
+		final BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
+		while (true) {
 
-			System.out.println("+++++ Processing cluster #" + count);
+			System.out.print("Enter minSupp:");
+			try {
+				minSupp = Double.parseDouble(br.readLine());
+			} catch (final NumberFormatException e) {
+				System.err.println("Invalid Format. Using " + minSupp);
+			}
+			if (minSupp <= 0)
+				break;
 
-			System.out.print("  Creating temporary transaction DB... ");
-			final File transactionDB = File.createTempFile("APICallDB", ".txt");
-			final BiMap<String, Integer> dictionary = HashBiMap.create();
-			generateTransactionDatabase(callSeqs, dictionary, transactionDB);
-			System.out.println("done.");
+			final File arffFileFreq = File.createTempFile("FreqCalls", ".arff");
+			writeArffHeader(arffFileFreq);
 
-			System.out.print("  Mining frequent sequences... ");
-			final File freqSeqs = File.createTempFile("APICallSeqs", ".txt");
-			FrequentSequenceMiner.mineFrequentClosedSequencesBIDE(transactionDB.getAbsolutePath(),
-					freqSeqs.getAbsolutePath(), minSupp);
-			transactionDB.delete();
-			System.out.println("done.");
+			int count = 0;
+			for (final Collection<String> callSeqs : clusteredCallSeqs1.asMap().values()) {
 
-			saveFrequentSequencesArffFile(freqSeqs, dictionary, arffFileFreq);
-			freqSeqs.delete();
+				System.out.println("+++++ Processing cluster #" + count);
 
-			count++;
-		}
+				System.out.print("  Creating temporary transaction DB... ");
+				final File transactionDB = File.createTempFile("APICallDB", ".txt");
+				final BiMap<String, Integer> dictionary = HashBiMap.create();
+				generateTransactionDatabase(callSeqs, dictionary, transactionDB);
+				System.out.println("done.");
 
-		System.out.print("===== Clustering call sequences #2... ");
-		final Multimap<Integer, String> clusteredCallSeqs2 = APICallClustererSequence
-				.clusterAPICallSeqs(arffFileFreq.getAbsolutePath(), noClusters2);
-		arffFileFreq.delete();
-		System.out.println("done. Number of clusters: " + clusteredCallSeqs2.keySet());
+				System.out.print("  Mining frequent sequences... ");
+				final File freqSeqs = File.createTempFile("APICallSeqs", ".txt");
+				FrequentSequenceMiner.mineFrequentClosedSequencesBIDE(transactionDB.getAbsolutePath(),
+						freqSeqs.getAbsolutePath(), minSupp);
+				transactionDB.delete();
+				System.out.println("done.");
 
-		count = 0;
-		for (final Collection<String> callSeqs : clusteredCallSeqs2.asMap().values()) {
-			final File outFile = new File(outFolder + "/Cluster" + count + "FreqCallSeqs.txt");
-			writeClusteredSequences(callSeqs, outFile);
-			count++;
+				saveFrequentSequencesArffFile(freqSeqs, dictionary, arffFileFreq);
+				freqSeqs.delete();
+
+				count++;
+			}
+
+			System.out.print("===== Clustering call sequences #2... ");
+			final Multimap<Integer, String> clusteredCallSeqs2 = APICallClustererSequence
+					.clusterAPICallSeqs(arffFileFreq.getAbsolutePath(), noClusters2);
+			arffFileFreq.delete();
+			System.out.println("done. Number of clusters: " + clusteredCallSeqs2.keySet());
+
+			count = 0;
+			for (final Collection<String> callSeqs : clusteredCallSeqs2.asMap().values()) {
+				final File outFile = new File(outFolder + "/Cluster" + count + "FreqCallSeqs.txt");
+				writeClusteredSequences(callSeqs, outFile);
+				count++;
+			}
 		}
 
 	}
